Change default locale to en
